requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Cors==5.0.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.2
werkzeug==3.0.4
zipp==3.20.2
